+++ frontend/src/page/training_session.rs
@@ -654,6 +654,7 @@ pub enum Msg {
     CreateExercise,
     ReplaceExercise(usize, usize, u32),
     DeferExercise(usize),
+    AddSet(usize),
     CloseDialog,
 
     UpdateStopwatchMetronomTimer,
@@ -1119,6 +1120,12 @@ pub fn update(
                 .send_msg(Msg::SaveTrainingSession)
                 .send_msg(Msg::CloseDialog);
         }
+        Msg::AddSet(section_idx) => {
+            add_set(&mut model.form.sections, section_idx);
+            orders
+                .send_msg(Msg::SaveTrainingSession)
+                .send_msg(Msg::CloseDialog);
+        }
         Msg::CloseDialog => {
             model.dialog = Dialog::Hidden;
         }
@@ -1386,6 +1393,112 @@ fn defer_exercise(sections: &mut [FormSection], section_idx: usize) {
     }
     sections[section_idx..section_idx + deferred_sections + preferred_sections]
         .rotate_right(preferred_sections);
+}
+
+fn add_set(sections: &mut Vec<FormSection>, section_idx: usize) {
+    if section_idx >= sections.len() {
+        return;
+    }
+
+    let ids = if let FormSection::Set { exercises } = &sections[section_idx] {
+        exercises.iter().map(|e| e.exercise_id).collect::<Vec<_>>()
+    } else {
+        return;
+    };
+
+    // Determine first set with same exercises
+    let mut first_idx = section_idx;
+
+    for (i, section) in sections
+        .iter()
+        .rev()
+        .enumerate()
+        .skip(sections.len() - section_idx + 1)
+    {
+        if let FormSection::Set { exercises } = &section {
+            let exercise_ids = exercises.iter().map(|e| e.exercise_id).collect::<Vec<_>>();
+            if ids == exercise_ids {
+                first_idx = section_idx - (sections.len() - i + 1);
+            } else {
+                break;
+            }
+        }
+    }
+
+    // Determine last set with same exercises
+    let mut last_idx = section_idx;
+
+    for (i, section) in sections.iter().enumerate().skip(section_idx + 1) {
+        if let FormSection::Set { exercises } = &section {
+            let exercise_ids = exercises.iter().map(|e| e.exercise_id).collect::<Vec<_>>();
+            if ids == exercise_ids {
+                last_idx = i;
+            } else {
+                break;
+            }
+        }
+    }
+
+    assert!(first_idx <= last_idx);
+    assert!(last_idx < sections.len());
+
+    // Determine rest between sets
+    let rest_idx = if section_idx < last_idx {
+        section_idx + 1
+    } else if first_idx < section_idx {
+        section_idx - 1
+    } else if section_idx + 1 < sections.len() {
+        section_idx + 1
+    } else {
+        section_idx
+    };
+
+    let rest = if let FormSection::Rest {
+        target_time,
+        automatic,
+    } = &sections[rest_idx]
+    {
+        FormSection::Rest {
+            target_time: *target_time,
+            automatic: *automatic,
+        }
+    } else {
+        FormSection::Rest {
+            target_time: 0,
+            automatic: true,
+        }
+    };
+
+    // Add rest and set
+    sections.insert(section_idx + 1, rest);
+
+    if let FormSection::Set { exercises } = &sections[section_idx] {
+        sections.insert(
+            section_idx + 2,
+            FormSection::Set {
+                exercises: exercises
+                    .iter()
+                    .map(|e| ExerciseForm {
+                        exercise_id: e.exercise_id,
+                        exercise_name: e.exercise_name.to_string(),
+                        reps: common::InputField::default(),
+                        time: common::InputField::default(),
+                        weight: common::InputField::default(),
+                        rpe: common::InputField::default(),
+                        target_reps: e.target_reps,
+                        target_time: e.target_time,
+                        target_weight: e.target_weight,
+                        target_rpe: e.target_rpe,
+                        prev_reps: None,
+                        prev_time: None,
+                        prev_weight: None,
+                        prev_rpe: None,
+                        automatic: e.automatic,
+                    })
+                    .collect::<Vec<_>>(),
+            },
+        );
+    }
 }
 
 // ------ ------
@@ -2126,7 +2239,23 @@ fn view_options_dialog(section_idx: usize, exercise_idx: usize) -> Vec<Node<Msg>
                     ]
                 ]
             ]
-        ]
+        ],
+        IF![exercise_idx == 0 =>
+            p![
+                C!["mt-3"],
+                a![
+                    C!["has-text-weight-bold"],
+                    ev(Ev::Click, move |_| Msg::AddSet(
+                        section_idx
+                    )),
+                    span![
+                        C!["icon-text"],
+                        span![C!["icon"], i![C!["fas fa-plus"]]],
+                        span!["Add set"],
+                    ]
+                ]
+            ]
+        ],
     ]
 }
 
@@ -2580,6 +2709,316 @@ mod tests {
                 rest(2),
                 set(vec![exercise(6, 0), exercise(7, 2)]),
                 rest(3),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_first_set() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            rest(0),
+            set(vec![exercise(1, 0)]),
+            rest(1),
+            set(vec![exercise(2, 1)]),
+            rest(2),
+            set(vec![exercise(3, 1)]),
+            rest(3),
+        ];
+        add_set(&mut sections, 0);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                rest(0),
+                set(vec![exercise(0, 0)]),
+                rest(0),
+                set(vec![exercise(1, 0)]),
+                rest(1),
+                set(vec![exercise(2, 1)]),
+                rest(2),
+                set(vec![exercise(3, 1)]),
+                rest(3),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_second_set() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            rest(0),
+            set(vec![exercise(1, 0)]),
+            rest(1),
+            set(vec![exercise(2, 1)]),
+            rest(2),
+            set(vec![exercise(3, 1)]),
+            rest(3),
+        ];
+        add_set(&mut sections, 2);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                rest(0),
+                set(vec![exercise(1, 0)]),
+                rest(0),
+                set(vec![exercise(1, 0)]),
+                rest(1),
+                set(vec![exercise(2, 1)]),
+                rest(2),
+                set(vec![exercise(3, 1)]),
+                rest(3),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_penultimate_set() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            rest(0),
+            set(vec![exercise(1, 0)]),
+            rest(1),
+            set(vec![exercise(2, 1)]),
+            rest(2),
+            set(vec![exercise(3, 1)]),
+            rest(3),
+        ];
+        add_set(&mut sections, 4);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                rest(0),
+                set(vec![exercise(1, 0)]),
+                rest(1),
+                set(vec![exercise(2, 1)]),
+                rest(2),
+                set(vec![exercise(2, 1)]),
+                rest(2),
+                set(vec![exercise(3, 1)]),
+                rest(3),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_last_set() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            rest(0),
+            set(vec![exercise(1, 0)]),
+            rest(1),
+            set(vec![exercise(2, 1)]),
+            rest(2),
+            set(vec![exercise(3, 1)]),
+            rest(3),
+        ];
+        add_set(&mut sections, 6);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                rest(0),
+                set(vec![exercise(1, 0)]),
+                rest(1),
+                set(vec![exercise(2, 1)]),
+                rest(2),
+                set(vec![exercise(3, 1)]),
+                rest(2),
+                set(vec![exercise(3, 1)]),
+                rest(3),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_superset() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0), exercise(4, 2)]),
+            rest(0),
+            set(vec![exercise(1, 0), exercise(5, 2)]),
+            rest(1),
+        ];
+        add_set(&mut sections, 0);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0), exercise(4, 2)]),
+                rest(0),
+                set(vec![exercise(0, 0), exercise(4, 2)]),
+                rest(0),
+                set(vec![exercise(1, 0), exercise(5, 2)]),
+                rest(1),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_no_rest_first_set() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            set(vec![exercise(1, 0)]),
+            set(vec![exercise(2, 1)]),
+            set(vec![exercise(3, 1)]),
+        ];
+        add_set(&mut sections, 0);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                rest(0),
+                set(vec![exercise(0, 0)]),
+                set(vec![exercise(1, 0)]),
+                set(vec![exercise(2, 1)]),
+                set(vec![exercise(3, 1)]),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_no_rest_second_set() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            set(vec![exercise(1, 0)]),
+            set(vec![exercise(2, 1)]),
+            set(vec![exercise(3, 1)]),
+        ];
+        add_set(&mut sections, 1);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                set(vec![exercise(1, 0)]),
+                rest(0),
+                set(vec![exercise(1, 0)]),
+                set(vec![exercise(2, 1)]),
+                set(vec![exercise(3, 1)]),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_no_rest_penultimate_set() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            set(vec![exercise(1, 0)]),
+            set(vec![exercise(2, 1)]),
+            set(vec![exercise(3, 1)]),
+        ];
+        add_set(&mut sections, 2);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                set(vec![exercise(1, 0)]),
+                set(vec![exercise(2, 1)]),
+                rest(0),
+                set(vec![exercise(2, 1)]),
+                set(vec![exercise(3, 1)]),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_no_rest_last_set() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            set(vec![exercise(1, 0)]),
+            set(vec![exercise(2, 1)]),
+            set(vec![exercise(3, 1)]),
+        ];
+        add_set(&mut sections, 3);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                set(vec![exercise(1, 0)]),
+                set(vec![exercise(2, 1)]),
+                set(vec![exercise(3, 1)]),
+                rest(0),
+                set(vec![exercise(3, 1)]),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_first_single_set() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            rest(0),
+            set(vec![exercise(1, 1)]),
+        ];
+        add_set(&mut sections, 0);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                rest(0),
+                set(vec![exercise(0, 0)]),
+                rest(0),
+                set(vec![exercise(1, 1)]),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_last_single_set() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            rest(0),
+            set(vec![exercise(1, 1)]),
+        ];
+        add_set(&mut sections, 2);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                rest(0),
+                set(vec![exercise(1, 1)]),
+                rest(0),
+                set(vec![exercise(1, 1)]),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_invalid_section_idx_rest() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            rest(0),
+            set(vec![exercise(1, 0)]),
+            rest(1),
+        ];
+        add_set(&mut sections, 1);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                rest(0),
+                set(vec![exercise(1, 0)]),
+                rest(1),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_add_set_invalid_section_idx_out_of_range() {
+        let mut sections = vec![
+            set(vec![exercise(0, 0)]),
+            rest(0),
+            set(vec![exercise(1, 0)]),
+            rest(1),
+        ];
+        add_set(&mut sections, 4);
+        assert_eq!(
+            sections,
+            vec![
+                set(vec![exercise(0, 0)]),
+                rest(0),
+                set(vec![exercise(1, 0)]),
+                rest(1),
             ]
         );
     }